feat(paste images): add images using `paste`
